pyproject.toml
[tool.poetry]
name = "lms-system"
version = "0.1.0"
description = ""
authors = ["Alexey Solovyev <solovaa@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
djangorestframework = "^3.16.1"
pillow = "^12.0.0"
django-filter = "^25.2"


[tool.poetry.group.lint.dependencies]
black = "^25.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
